trunk: corrected integer kind

+++ src/base/dataio.F90
@@ -995,7 +995,7 @@ contains
 
       type(component_fluid), intent(inout), target :: fl
       type(phys_prop), pointer                     :: pr
-      integer(kind=4)                              :: wa_i
+      integer                                      :: wa_i
       type(cg_list_element), pointer               :: cgl
 
       wa_i = all_cg%ind(wa_n)
@@ -1118,7 +1118,7 @@ contains
 
       real                           :: dxmn_safe
       type(cg_list_element), pointer :: cgl
-      integer(kind=4)                :: wa_i
+      integer                        :: wa_i
       type(value)                    :: drag
 #ifdef MAGNETIC
       type(value)                    :: b_min, b_max, divb_max, vai_max
